Code cleanups
This touches nearly all files, but only few things have changed:

 - Consistent use of EOL for all source files
 - Returning nullptr instead of 0 in MemWatchDelegate::createEditor
 - Add // namespace X after the closing brace for namespaces
 - Fix warning in DlgAddWatchEntry (casting size_t to int)

+++ Source/GUI/MemWatcher/DlgAddWatchEntry.cpp
@@ -189,7 +189,7 @@ DlgAddWatchEntry::~DlgAddWatchEntry()
 
 void DlgAddWatchEntry::addPointerOffset()
 {
-  int level = m_entry->getPointerLevel();
+  int level = static_cast<int>(m_entry->getPointerLevel());
   QLabel* lblLevel = new QLabel(QString::fromStdString("Level " + std::to_string(level + 1) + ":"));
   QLineEdit* txbOffset = new QLineEdit();
   txbOffset->setText(0);
@@ -210,7 +210,7 @@ void DlgAddWatchEntry::addPointerOffset()
 
 void DlgAddWatchEntry::removePointerOffset()
 {
-  int level = m_entry->getPointerLevel();
+  int level = static_cast<int>(m_entry->getPointerLevel());
 
   QLayoutItem* lblLevelItem = m_offsetsLayout->takeAt(
       m_offsetsLayout->indexOf(m_offsetsLayout->itemAtPosition(level - 1, 0)->widget()));
@@ -364,7 +364,7 @@ void DlgAddWatchEntry::updatePreview()
   m_lblValuePreview->setText(QString::fromStdString(m_entry->getStringFromMemory()));
   if (m_entry->isBoundToPointer())
   {
-    int level = m_entry->getPointerLevel();
+    size_t level = m_entry->getPointerLevel();
     for (int i = 0; i < level; ++i)
     {
       QWidget* test = m_offsetsLayout->itemAtPosition(i, 2)->widget();
